+++ lib/src/model/app_info_model.dart
@@ -3,7 +3,6 @@ import 'package:flutter/material.dart' show Colors;
 import 'package:flutter_content/flutter_content.dart';
 import 'package:flutter_content/src/api/snippet_panel/clipboard_view.dart';
 import 'package:flutter_content/src/snippet/pnodes/groups/button_style_properties.dart';
-import 'package:flutter_content/src/snippet/pnodes/groups/container_style_properties.dart';
 import 'package:flutter_content/src/snippet/pnodes/groups/text_style_properties.dart';
 
 part 'app_info_model.mapper.dart';