+++ lib/src/snippet/pnodes/container_style_pnodes.dart
@@ -8,7 +8,6 @@ import 'package:flutter_content/src/snippet/pnodes/enum_pnode.dart';
 import 'package:flutter_content/src/snippet/pnodes/enums/mappable_enum_decoration.dart';
 import 'package:flutter_content/src/snippet/pnodes/fyi_pnodes.dart';
 import 'package:flutter_content/src/snippet/pnodes/gradient_pnode.dart';
-import 'package:flutter_content/src/snippet/pnodes/groups/container_style_properties.dart';
 import 'package:flutter_content/src/snippet/pnodes/size_pnode.dart';
 import 'package:flutter_content/src/text_styles/style_name_editor.dart';