+++ lib/x_flutter_content/google_font_names_extn.dart
@@ -1,5 +1,4 @@
 import 'package:flutter_content/flutter_content.dart';
-import 'package:flutter_content/src/snippet/pnodes/groups/text_style_properties.dart';
 
 extension GoogleFontNamesExtension on FlutterContentMixins {
   void loadGoogleFontNames(List<GoogleFontName> arr) => arr.addAll(const [