+++ CHANGELOG.md
@@ -4,9 +4,9 @@ All notable changes to this project will be documented in this file.
 The message format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## 3.10.0 - 2025-08-02
-### Changed
-- snippet triangle now has menu
+## 3.11.0 - 2025-08-05
+### Fixed
+- fix EditablePage s.t. SnippetBeingEdited rendered via a SnippetPanel
 
 ## 5.0.0 - 2025-08-02
 ### Fixed
@@ -16,6 +16,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## 4.0.0 - 2025-08-02
 ### Added
 - no longer need to initWithContext + other fca improvements
+
+## 3.10.0 - 2025-08-02
+### Changed
+- snippet triangle now has menu
 
 ## 3.9.0 - 2025-08-02
 ### Added